+++ src/camping/actor.ts
@@ -4,24 +4,40 @@ import {capitalize, unslugify} from '../utils';
 
 type ItemType = 'effect' | 'consumable';
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
 async function getSourceIds(uuids: Set<string>): Promise<Set<string>> {
-    const applicableItems = await Promise.all(Array.from(uuids).map(uuid => {
-        return fromUuid(uuid);
-    }));
+    const applicableItems = (await Promise.all(Array.from(uuids).map(uuid => fromUuid(uuid))) as (Item | null)[])
+        .filter(i => i !== null) as Item[];
     const sourceIds = applicableItems
         .filter(item => item !== undefined && item !== null)
-        .map((item: any) => item.sourceId);
+        .map(item => item.sourceId);
     return new Set(sourceIds);
 }
 
 export function getItemsBySourceId(actor: Actor, type: ItemType, sourceIds: Set<string>): Item[] {
     return actor.itemTypes[type]
-        .filter((a: any) => sourceIds.has(a.sourceId));
+        .filter(i => sourceIds.has(i.sourceId));
+}
+
+export function isEffectItem(item: Item): item is Item & EffectItem {
+    return item.type === 'effect';
+}
+
+export function isConsumableItem(item: Item): item is Item & ConsumableItem {
+    return item.type === 'consumable';
 }
 
 export async function getItemsByUuid(actor: Actor, type: ItemType, uuids: Set<string>): Promise<Item[]> {
     return getItemsBySourceId(actor, type, await getSourceIds(uuids));
+}
+
+export async function getEffectsByUuid(actor: Actor, uuids: Set<string>): Promise<(Item & EffectItem)[]> {
+    const items = await getItemsByUuid(actor, 'effect', uuids);
+    return items.filter(isEffectItem);
+}
+
+export async function getConsumablesByUuid(actor: Actor, uuids: Set<string>): Promise<(Item & ConsumableItem)[]> {
+    const items = await getItemsByUuid(actor, 'consumable', uuids);
+    return items.filter(isConsumableItem);
 }
 
 export async function hasItemByUuid(actor: Actor, type: ItemType, uuids: Set<string>): Promise<boolean> {
@@ -39,7 +55,8 @@ export async function removeItemsBySourceId(actors: Actor[], type: ItemType, app
 export async function removeExpiredEffects(actors: Actor[], applicableSourceIds: Set<string>): Promise<void> {
     for (const actor of actors) {
         const expiredEffectIds = getItemsBySourceId(actor, 'effect', applicableSourceIds)
-            .filter((a: any) => a.isExpired)
+            .filter(isEffectItem)
+            .filter(a => a.isExpired)
             .map(a => a.id) as string[];
         await actor.deleteEmbeddedDocuments('Item', expiredEffectIds);
     }
@@ -75,11 +92,14 @@ export async function validateSkillProficiencies(actor: Actor, requirements: Ski
 export async function satisfiesSkillProficiency(actor: Actor, skill: string, minimumProficiency: Proficiency): Promise<boolean> {
     const skillProp = await getSkill(actor, skill);
     const rank = proficiencyToRank(minimumProficiency);
-    return skillProp?.rank >= rank ?? false;
+    if (skillProp) {
+        return skillProp.rank >= rank;
+    }
+    return false;
 }
 
-export async function getSkill(actor: Actor, skill: string): Promise<any | null> {
-    const skills = (actor as any).skills;
+export async function getSkill(actor: Actor, skill: string): Promise<ActorSkill | null> {
+    const skills = actor.skills;
     const loreSkill = `${skill}-lore`;
     if (skill in skills) {
         return skills[skill];
@@ -91,9 +111,9 @@ export async function getSkill(actor: Actor, skill: string): Promise<any | null>
 }
 
 export async function hasCookingLore(actorUuid: string): Promise<boolean> {
-    const actor = await fromUuid(actorUuid);
+    const actor = await fromUuid(actorUuid) as Actor | null;
     if (actor) {
-        return await getSkill((actor as any), 'cooking') !== null;
+        return await getSkill(actor, 'cooking') !== null;
     } else {
         return false;
     }