+++ src/skill-checks.ts
@@ -5,9 +5,7 @@ async function rollSkillCheck(
 ): Promise<void> {
     const checkDc = dc === undefined ? undefined : {value: dc};
     for (const actor of actors) {
-        /* eslint-disable @typescript-eslint/no-explicit-any */
-        const pf2eActor = actor as any;
-        const check = skill === 'perception' ? pf2eActor.perception : pf2eActor.skills[skill];
+        const check = skill === 'perception' ? actor.perception : actor.skills[skill];
         await check.roll({
             rollMode: 'gmroll',
             dc: checkDc,